user role membership of deleted user roles

+++ workflix/backend/backend/src/main/kotlin/de/se/team3/logic/container/UserRoleMembershipContainer.kt
@@ -10,12 +10,15 @@ object UserRoleMembershipContainer : UserRoleMembershipContainerInfterface {
     /**
      * Creates the specified membership of a user in a user role.
      *
+     * @throws NotFoundException Is thrown if the specified user role does not exist.
      * @throws AlreadyExistsException Is thrown if the specified membership already exists.
      */
     override fun addUserToRole(userID: String, userRoleID: Int) {
         val userRole = UserRoleContainer.getUserRole(userRoleID) // throws not found exception
         val user = UserContainer.getUser(userID) // throws NotFoundException
 
+        if (userRole.isDeleted())
+            throw NotFoundException("user role does not exist")
         if (userRole.hasMember(user.id))
             throw AlreadyExistsException("the membership already exists")
 
@@ -27,11 +30,15 @@ object UserRoleMembershipContainer : UserRoleMembershipContainerInfterface {
     /**
      * Deletes the specified membership of a user in a user role.
      *
+     * @throws NotFoundException Is thrown if the specified user role does not exist.
      * @throws NotFoundException Is thrown if the specified membership does not exist.
      */
     override fun deleteUserFromRole(userID: String, userRoleID: Int) {
         val userRole = UserRoleContainer.getUserRole(userRoleID)
         val user = UserContainer.getUser(userID)
+
+        if (userRole.isDeleted())
+            throw NotFoundException("user role does not exist")
 
         val existed = UserRoleMembershipDAO.deleteUserFromRole(userID, userRoleID)
         if (!existed)
